Split account mutations for files
